LPD-53896 Rename, use internal name as context is not exposed outside FDS
